chore(scavenge): remove in-memory backend
