--- src/EventStore.Core/TransactionLog/Scavenging/InMemory/InMemoryTransactionManager.cs
@@ -1,4 +1,0 @@
-namespace EventStore.Core.TransactionLog.Scavenging.InMemory;
-
-public class InMemoryTransactionManager(InMemoryScavengeMap<Unit, ScavengeCheckpoint> storage)
-	: TransactionManager<int>(new InMemoryTransactionFactory(), storage);
